// 07-Pokedex/src/interfaces/IPokemon.tsx
export interface SimplePokemon {
    id: string;
    name: string;
    picture:  string;
    color?: string;
}